+++ frontend/src/components/review/state/review-atoms.ts
@@ -1,6 +1,7 @@
 import { Term } from "gql/codegen-output";
 import { atom } from "recoil";
-import { ReviewStage } from "../types/review.types";
+import { ReviewSessionWithoutUserIdInput } from "../../../gql/codegen-output";
+import { ReviewStage, SessionEntryWithoutTimeOnCard } from "../types/review.types";
 
 export const reviewStageState = atom<ReviewStage>({
 	key: "reviewStageState",
@@ -10,4 +11,14 @@ export const reviewStageState = atom<ReviewStage>({
 export const reviewCardTermsState = atom<Term[]>({
 	default: [],
 	key: "reviewCardTerms",
+});
+
+export const reviewSessionState = atom<Partial<ReviewSessionWithoutUserIdInput>>({
+	default: {},
+	key: "reviewSessionState",
+});
+
+export const reviewEntriesState = atom<Partial<SessionEntryWithoutTimeOnCard>[]>({
+	default: [],
+	key: "reviewEntries",
 });
